check shard status before mutating shard state

+++ adapters/repos/db/shard_geo_props.go
@@ -78,6 +78,10 @@ func geoPropID(shardID string, propName string) string {
 
 func (s *Shard) updatePropertySpecificIndices(object *storobj.Object,
 	status objectInsertStatus) error {
+	if s.isReadOnly() {
+		return ErrShardReadOnly
+	}
+
 	for propName, propIndex := range s.propertyIndices {
 		if err := s.updatePropertySpecificIndex(propName, propIndex,
 			object, status); err != nil {
@@ -101,6 +105,10 @@ func (s *Shard) updatePropertySpecificIndex(propName string,
 
 func (s *Shard) updateGeoIndex(propName string, index propertyspecific.Index,
 	obj *storobj.Object, status objectInsertStatus) error {
+	if s.isReadOnly() {
+		return ErrShardReadOnly
+	}
+
 	if status.docIDChanged {
 		if err := s.deleteFromGeoIndex(index, status.oldDocID); err != nil {
 			return errors.Wrap(err, "delete old doc id from geo index")
@@ -112,6 +120,10 @@ func (s *Shard) updateGeoIndex(propName string, index propertyspecific.Index,
 
 func (s *Shard) addToGeoIndex(propName string, index propertyspecific.Index,
 	obj *storobj.Object, status objectInsertStatus) error {
+	if s.isReadOnly() {
+		return ErrShardReadOnly
+	}
+
 	if obj.Properties() == nil {
 		return nil
 	}
@@ -138,6 +150,10 @@ func (s *Shard) addToGeoIndex(propName string, index propertyspecific.Index,
 
 func (s *Shard) deleteFromGeoIndex(index propertyspecific.Index,
 	docID uint64) error {
+	if s.isReadOnly() {
+		return ErrShardReadOnly
+	}
+
 	if err := index.GeoIndex.Delete(docID); err != nil {
 		return errors.Wrapf(err, "delete from geo index")
 	}
